requirements.txt
streamlit
pandas
matplotlib
seaborn
langchain-groq
langchain
python-dotenv